Extract Lookback to the Cycles namespace
